refactor(extension): Rename kernel dir to kernel-integration
