+++ packages/extension/src/kernel-integration/VatWorkerClient.test.ts
