+++ packages/extension/src/kernel-integration/VatWorkerClient.ts
