+++ packages/extension/src/kernel-integration/handle-panel-message.test.ts
