+++ packages/extension/src/kernel-integration/iframe-vat-worker.ts
