+++ packages/extension/src/kernel-integration/kernel-worker.ts
