+++ packages/extension/src/kernel-integration/run-vat-lifecycle.test.ts
