+++ packages/extension/src/kernel-integration/run-vat-lifecycle.ts
