+++ packages/extension/src/kernel-integration/vat-worker-service.test.ts
