+++ packages/extension/src/kernel-integration/vat-worker-service.ts
