+++ packages/extension/vite.config.ts
@@ -38,7 +38,10 @@ export default defineConfig(({ mode }) => ({
     rollupOptions: {
       input: {
         background: path.resolve(sourceDir, 'background.ts'),
-        'kernel-worker': path.resolve(sourceDir, 'kernel/kernel-worker.ts'),
+        'kernel-worker': path.resolve(
+          sourceDir,
+          'kernel-integration/kernel-worker.ts',
+        ),
         offscreen: path.resolve(sourceDir, 'offscreen.html'),
         iframe: path.resolve(sourceDir, 'iframe.html'),
         popup: path.resolve(sourceDir, 'popup.html'),
